+++ apps/alloy_dendrite_elastic/src/cpu/alloy_dendrite_planar.cpp
@@ -40,7 +40,7 @@
 #include <openpfc/runtime/common/mpi_main.hpp>
 
 #include <alloy_dendrite/cases.hpp>
-#include <alloy_dendrite/cli.hpp>
+#include <openpfc/frontend/utils/cli_options.hpp>
 
 namespace {
 
@@ -106,7 +106,7 @@ void print_usage(std::ostream &os, const char *exe) {
 }
 
 int run(int argc, char **argv, int rank, int nproc) {
-  alloy_dendrite::Options opt(argc, argv);
+  pfc::utils::CliOptions opt(argc, argv);
   if (opt.help()) {
     if (rank == 0) {
       print_usage(std::cout, argv[0]);